requirements.txt
Django>=2.2.5
django-debug-toolbar>=2.0
django-extensions>=2.2.1
djangorestframework>=3.10.3
pytz>=2019.2
six>=1.12.0
sqlparse>=0.3.0
Werkzeug>=0.16.0
